% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPNList.R
\name{BPNList}
\alias{BPNList}
\title{Construct a BPNList object}
\usage{
BPNList(input, gsea, enrichr, ilincs, leading, revigo, integrated)
}
\arguments{
\item{input}{tbl_df A tibble with the differentialy expressed genes}

\item{gsea}{GSEAResult An object containing the GSEA results}

\item{enrichr}{EnrichRResult An object containing the EnrichR results}

\item{ilincs}{iLINCSResult An object containing the iLINCS results}

\item{leading}{LEResult An object containing the Leading Edge Analysis results}

\item{revigo}{RevigoResult An object containing the Revigo results}

\item{integrated}{IntegratedResult An object containing the Integrated results}
}
\value{
A BPNList object
}
\description{
Construct a BPNList object
}
\examples{
TRUE
}
