% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_enrichr_pod.R
\name{do_enrichr_pod}
\alias{do_enrichr_pod}
\title{Perform the EnrichR Analysis}
\usage{
do_enrichr_pod(bpn, alpha = 0.05, lower = NULL, upper = NULL)
}
\arguments{
\item{bpn}{BPNList An object of class BPNList created from \link{prepare_data}}

\item{alpha}{numeric The Alpha value for statistical significance. Defaults to 0.05.}

\item{lower}{numeric The cutoff for down-regulated genes. If not specified, it takes the bottom 10% of the}

\item{upper}{numeric The cutoff for up-regulated genes. If not specified, it takes the top 10% of the dataset}
}
\value{
An object of class BPNList with its enrichr slot populated with the results
}
\description{
This function takes in a BPNList object and returns a BPNList object with the
enrichr slot updated with process EnrichR Results
}
\examples{
TRUE
}
