% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gsea_pod.R
\name{do_gsea_pod}
\alias{do_gsea_pod}
\title{Perform the Geneset Enrichment Analysis}
\usage{
do_gsea_pod(
  bpn,
  species = "hsapiens",
  gmtfile = NULL,
  lower = NULL,
  upper = NULL,
  alpha = 0.05,
  minSize = 15,
  maxSize = 500,
  rankWithSignificance = FALSE,
  useFDR = TRUE
)
}
\arguments{
\item{bpn}{BPNList An object of class BPNList created from \link{prepare_data}}

\item{species}{character One of the species for which we provide current GMT pathways. Currently the valid species are hsapiens, mmusculus and rnorvegicus. Cannot be specified along with gmtfile}

\item{gmtfile}{character A full path to a GMT File you would like to use. Cannot be specified along with species}

\item{lower}{numeric The cutoff for down-regulated genes. If not specified, it takes the bottom 10\% of the dataset}

\item{upper}{numeric The cutoff for up-regulated genes. If not specified, it takes the top 10\% of the dataset}

\item{alpha}{numeric The Alpha value for statistical significance. Defaults to 0.05.}

\item{minSize}{numeric The minimum number of genes that should be in a pathway for it to be included in the analysis}

\item{maxSize}{numeric The maximum number of genes that should be in a pathway for it to be included in the analysis}

\item{rankWithSignificance}{logical Whether to rank genes using p-values}

\item{useFDR}{logical Whether to filter by adjusted p-value}
}
\value{
An object of class BPNList with its gsea slot populated with the results
}
\description{
This function takes a BPNList object and performs a Geneset Enrichment Analysis
and returns a BPNList object with the results of the GSEA
}
\details{
This function takes a BPNList object and performs a Geneset Enrichment Analysis
and returns a BPNList object with the results of the GSEA
}
\examples{
TRUE
}
