% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_leading_pod.R
\name{do_leading_pod}
\alias{do_leading_pod}
\title{Perform Leading Edge Gene Analysis}
\usage{
do_leading_pod(bpn)
}
\arguments{
\item{bpn}{An object of class BPNList. Must have the gsea analysis completed beforehand.}
}
\value{
An object of class BPNList with its leading slot populated with the results
}
\description{
This method takes in a bpn object and performs a leading edge analysis to
identify the genes that are driving the majority of pathways.
}
\examples{
TRUE
}
