% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichRResult.R
\name{EnrichRResult}
\alias{EnrichRResult}
\title{Construct an EnrichRResult object}
\usage{
EnrichRResult(
  up_results,
  down_results,
  upreg,
  downreg,
  alpha,
  upper,
  lower,
  dbs
)
}
\arguments{
\item{up_results}{tbl_df A tibble of enrichment results from the upregulated genes}

\item{down_results}{tbl_df A tibble of enrichment results from the upregulated genes}

\item{upreg}{tbl_df A tibble of up-regulated genes}

\item{downreg}{tbl_df A tibble of down-regulated genes}

\item{alpha}{numeric Alpha level of significance}

\item{upper}{numeric Upper threshold for up-regulated genes}

\item{lower}{numeric Lower threshold for down-regulated genes}

\item{dbs}{character a vector of the databases used for the enrichment analysis}
}
\value{
an object of class EnrichRResult
}
\description{
Construct an EnrichRResult object
}
\examples{
TRUE
}
