% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_revigo_output.R
\name{get_revigo_output}
\alias{get_revigo_output}
\title{Perform Revigo enrichment analysis}
\usage{
get_revigo_output(input_data)
}
\arguments{
\item{input_data}{A dataframe containing GO IDs and associated p-values}
}
\value{
A dataframe containing the Revigo Enrichment analysis
}
\description{
Perform Revigo enrichment analysis
}
\examples{
TRUE
}
