% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAResult.R
\name{GSEAResult}
\alias{GSEAResult}
\title{Construct a GSEAResult Object}
\usage{
GSEAResult(results, pathways, lower, upper, alpha, upreg, downreg)
}
\arguments{
\item{results}{tbl_df A tibble with the results from fgsea}

\item{pathways}{list A list of vectors with gene sets}

\item{lower}{numeric Lower threshold for down-regulated genes}

\item{upper}{numeric Upper threshold for up-regulated genes}

\item{alpha}{numeric Alpha level of significance}

\item{upreg}{tbl_df A tibble of up-regulated genes}

\item{downreg}{tbl_df A tibble of down-regulated genes}
}
\value{
an object of class GSEAResult
}
\description{
Construct a GSEAResult Object
}
\examples{
TRUE
}
