% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_leading_pod.R
\name{leading_edge}
\alias{leading_edge}
\title{Extract Leading Edge Genes from GSEA Results}
\usage{
leading_edge(gsea, direction)
}
\arguments{
\item{gsea}{an object of class GSEAResult}

\item{direction}{one of "up" or "down"}
}
\value{
A tibble with the leading edge genes and associated pathways
}
\description{
Extract Leading Edge Genes from GSEA Results
}
\examples{
TRUE
}
