% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{LEResult-class}
\alias{LEResult-class}
\title{LEResult-class: Class containing Leading Edge Analysis results}
\value{
an object of class LEResult
}
\description{
LEResult-class: Class containing Leading Edge Analysis results
}
\section{Slots}{

\describe{
\item{\code{placeholder}}{numeric.}

\item{\code{empty}}{logical. A simple check for whether or not the object is empty}
}}

\examples{
TRUE
}
