% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEResult.R
\name{LEResult}
\alias{LEResult}
\title{Construct an LEResult object}
\usage{
LEResult(leading_up, leading_down)
}
\arguments{
\item{leading_up}{tbl_df. A dataframe with the results of all the leading edge genes from upregulated genes}

\item{leading_down}{tbl_df. A dataframe with the results of all the leading edge genes from downregulated genes}
}
\value{
An object of class LEResult
}
\description{
Construct an LEResult object
}
\examples{
TRUE
}
