% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gsea_pod.R
\name{load_pathways}
\alias{load_pathways}
\title{Load Pathways data}
\usage{
load_pathways(species)
}
\arguments{
\item{species}{The species. Can be hsapiens,  mmusculus or rnorvegicus}
}
\value{
The loaded pathways

importFrom("utils", "data")
}
\description{
Load Pathways data
}
\examples{
TRUE
}
