% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmusculus}
\alias{mmusculus}
\title{List of pathways and associated genes in mus musculus}
\format{
A list
}
\source{
\url{http://download.baderlab.org/EM_Genesets/October_01_2020/Mouse/symbol/Mouse_GO_AllPathways_with_GO_iea_October_01_2020_symbol.gmt}
}
\usage{
mmusculus
}
\description{
The processed form of the October 2020 Pathway annotation
containing every pathway annotated with the associated
genes in mice. This uses the Symbol annotation from BaderLab.
}
\keyword{datasets}
