% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare the data for downstream analysis}
\usage{
prepare_data(genes, logfc, pvalues = NULL)
}
\arguments{
\item{genes}{a character vector of gene names}

\item{logfc}{a numeric vector of log fold-change values}

\item{pvalues}{(optional) a numeric vector of p-values}
}
\value{
An object of class BPNList with the following slot set:

input
}
\description{
This function takes vectors of gene names, log Fold-change values and optionally p-values
and formats them into a tibble for downstream analysis.
}
\details{
The function returns an object of class BPNList that will update as the analysis proceeds
}
\examples{
TRUE
}
