% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{RevigoResult-class}
\alias{RevigoResult-class}
\title{RevigoResult-class: Class containing results from Revigo}
\value{
and object of class RevigoResult
}
\description{
RevigoResult-class: Class containing results from Revigo
}
\section{Slots}{

\describe{
\item{\code{placeholder}}{numeric.}

\item{\code{empty}}{logical. A simple check for whether or not the object is empty}
}}

\examples{
TRUE
}
