% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnorvegicus}
\alias{rnorvegicus}
\title{List of pathways and associated genes in rattus norvegicus}
\format{
A list
}
\source{
\url{http://download.baderlab.org/EM_Genesets/October_01_2020/MRat/symbol/Rat_GO_AllPathways_with_GO_iea_October_01_2020_symbol.gmt}
}
\usage{
rnorvegicus
}
\description{
The processed form of the October 2020 Pathway annotation
containing every pathway annotated with the associated
genes in rats. This uses the Symbol annotation from BaderLab.
}
\keyword{datasets}
