% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{IntegratedResult-class}
\alias{IntegratedResult-class}
\title{IntegratedResult-class: Class containing Integrated results}
\value{

}
\description{
IntegratedResult-class: Class containing Integrated results
}
\section{Slots}{

\describe{
\item{\code{placeholder}}{numeric.}
}}

\examples{
TRUE
}
