% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{LEResult-class}
\alias{LEResult-class}
\title{LEResult-class: Class containing Leading Edge Analysis results}
\value{
an object of class LEResult
}
\description{
LEResult-class: Class containing Leading Edge Analysis results
}
\section{Slots}{

\describe{
\item{\code{leading_up}}{tbl_df. A tibble with the results of LE analysis for the upregulated pathways}

\item{\code{leading_down}}{tbl_df. A tibble with the results of LE analysis for the downregulated pathways}

\item{\code{count_leading_up}}{tbl_df. A tibble with the counts of LE genes for the upregulated pathways}

\item{\code{count_leading_down}}{tbl_df. A tibble with the counts of LE genes for the downregulated pathways}

\item{\code{num_leading_up}}{numeric. Number of LE genes for the upregulated pathways}

\item{\code{num_leading_down}}{numeric. Number of LE genes for the downregulated pathways}

\item{\code{empty}}{logical. A simple check for whether or not the object is empty}
}}

\examples{
TRUE
}
